Add support for building templates on Proxmox
Added support for building images and using them as templates on the
Proxmox VE platform using the built-in provider.

The resulting template has cloud-init support enabled and uses a custom
partitioning scheme with simple file system layout allowing for it to be
easily resized and manipulated on startup through cloud-init.

Additionally, a number of small updates:

  - Removed the deprecated `spot_price_auto_product` option
    from the AWS EBS builder template

  - Replaced now deprecated `ssh_wait_timeout` option with
    currently supported `ssh_timeout` where applicable

  - Replaced now defunct Docker repository that offered
    legacy Docker packages with a mirror from Yandex,
    so that Docker can still be installed on Ubuntu 12.04

  - Removed Ubuntu Extended Security Maintenance (ESM)
    commercial repository that will now be added when
    building Ubuntu 14.04 images

  - Fixed issue with Python's pip when run under the
    `virtualenv` that has been installed via pip to be
    the latest version available. The `virtualenv` has
    has been pinned to version 16.7.10 on Ubuntu 12.04
    and 14.04, as the latest version 20.0.x would no
    longer install and work correctly

  - Connectivity test to check whether https://2ton.com.au/
    could be reached will no longer leave a stray `index.html`
    file in current user home directory

  - Moved to XZ compression for the initramfs so that the
    resulting file size would be smaller

  - Changed LVM partitioning scheme for the Vagrant boxes,
    so that the logical volume for swap space is no longer
    created automatically

+++ scripts/vagrant/minimize.sh
@@ -8,10 +8,12 @@ readonly SWAP_UUID=$(blkid -o value -l -s UUID -t TYPE=swap)
 readonly SWAP_PARTITION=$(readlink -f "/dev/disk/by-uuid/${SWAP_UUID}")
 
 # Zero the swap partition and re-initialize.
-swapoff "$SWAP_PARTITION"
-dd if=/dev/zero of="${SWAP_PARTITION}" bs=1M || true
-mkswap -U "$SWAP_UUID" "$SWAP_PARTITION"
-sync
+if [[ -n $SWAP_UUID ]]; then
+    swapoff "$SWAP_PARTITION"
+    dd if=/dev/zero of="${SWAP_PARTITION}" bs=1M || true
+    mkswap -U "$SWAP_UUID" "$SWAP_PARTITION"
+    sync
+fi
 
 # Zero the root partition to reclaim deleted space (allocations). This
 # will enable the resulting image to have better compression ratio.
